extracting hooks in useCausalDiscoveryRunner
