+++ javascript/app-discover/src/state/hooks/useDerivedConstraints.ts
@@ -1,0 +1,33 @@
+/*!
+ * Copyright (c) Microsoft. All rights reserved.
+ * Licensed under the MIT license. See LICENSE file in the project.
+ */
+import { useMemo } from 'react'
+
+import type { CausalVariable } from '../../domain/CausalVariable.js'
+import type { RelationshipReference } from '../../domain/Relationship.js'
+
+export function useDerivedConstraints(
+	inModelCausalVariables: CausalVariable[],
+): RelationshipReference[] {
+	return useMemo<RelationshipReference[]>(() => {
+		const result: RelationshipReference[] = []
+		inModelCausalVariables.forEach(sourceVar => {
+			sourceVar.derivedFrom?.forEach(sourceColumn => {
+				inModelCausalVariables.forEach(targetVar => {
+					if (
+						sourceVar !== targetVar &&
+						(targetVar.derivedFrom?.includes(sourceColumn) ||
+							sourceVar.disallowedRelationships?.includes(targetVar.columnName))
+					) {
+						result.push({
+							source: sourceVar,
+							target: targetVar,
+						})
+					}
+				})
+			})
+		})
+		return result
+	}, [inModelCausalVariables])
+}
