+++ javascript/app-discover/src/state/hooks/useLastDiscoveryResultPromise.ts
@@ -6,9 +6,14 @@ import { useCallback, useRef } from 'react'
 
 import type { CausalDiscoveryResultPromise } from '../../domain/CausalDiscovery/CausalDiscoveryResult.js'
 
+export type SetLastDiscoveryResultPromise = (
+	newPromise?: CausalDiscoveryResultPromise,
+) => void
+export type CancelLastDiscoveryResultPromise = () => Promise<void>
+
 export function useLastDiscoveryResultPromise(): [
-	(newPromise?: CausalDiscoveryResultPromise) => void,
-	() => Promise<void>,
+	SetLastDiscoveryResultPromise,
+	CancelLastDiscoveryResultPromise,
 ] {
 	const lastDiscoveryResultPromise = useRef<
 		CausalDiscoveryResultPromise | undefined
